refactor: 쿠폰 리팩토링
